Added cancelation to URLSession tasks
