+++ Sources/ApexyURLSession/URLSession+AsyncAwait13.swift
@@ -5,6 +5,7 @@
 //  Created by Aleksei Tiurnin on 20.01.2022.
 //
 
+import Apexy
 import Foundation
 
 extension URLSession {
@@ -15,19 +16,26 @@ extension URLSession {
     func data(
         for request: URLRequest,
         delegate: URLSessionTaskDelegate? = nil) async throws -> (Data, URLResponse) {
-        try await withCheckedThrowingContinuation { continuation in
-            let task = self.dataTask(with: request) { data, response, error in
-                guard let data = data, let response = response else {
-                    let error = error ?? URLError(.badServerResponse)
-                    return continuation.resume(throwing: error)
+            typealias ContentContinuation = CheckedContinuation<(Data, URLResponse), Error>
+            let progressWrapper = ProgressWrapper()
+            return try await withTaskCancellationHandler(handler: {
+                progressWrapper.cancel()
+            }, operation: {
+                try await withCheckedThrowingContinuation { (continuation: ContentContinuation) in
+                    let task = self.dataTask(with: request) { data, response, error in
+                        guard let data = data, let response = response else {
+                            let error = error ?? URLError(.badServerResponse)
+                            return continuation.resume(throwing: error)
+                        }
+                        
+                        continuation.resume(returning: (data, response))
+                    }
+                    progressWrapper.progress = task.progress
+                    
+                    task.resume()
                 }
-                
-                continuation.resume(returning: (data, response))
-            }
-            
-            task.resume()
+            })
         }
-    }
     
     @available(macOS, introduced: 10.15, deprecated: 12, message: "Extension is no longer necessary. Use API built into SDK")
     @available(iOS, introduced: 13, deprecated: 15, message: "Extension is no longer necessary. Use API built into SDK")
@@ -37,18 +45,27 @@ extension URLSession {
         for request: URLRequest,
         fromFile fileURL: URL,
         delegate: URLSessionTaskDelegate? = nil) async throws -> (Data, URLResponse) {
-        try await withCheckedThrowingContinuation { continuation in
-            let task = self.uploadTask(with: request, fromFile: fileURL) { data, response, error in
-                guard let data = data, let response = response else {
-                    let error = error ?? URLError(.badServerResponse)
-                    return continuation.resume(throwing: error)
+            typealias ContentContinuation = CheckedContinuation<(Data, URLResponse), Error>
+            let progressWrapper = ProgressWrapper()
+            return try await withTaskCancellationHandler(handler: {
+                progressWrapper.cancel()
+            }, operation: {
+                try await withCheckedThrowingContinuation { (continuation: ContentContinuation) in
+                    let task = self.uploadTask(with: request, fromFile: fileURL) { data, response, error in
+                        guard let data = data, let response = response else {
+                            let error = error ?? URLError(.badServerResponse)
+                            return continuation.resume(throwing: error)
+                        }
+                        
+                        continuation.resume(returning: (data, response))
+                    }
+                    
+                    progressWrapper.progress = task.progress
+                    
+                    task.resume()
                 }
-                
-                continuation.resume(returning: (data, response))
-            }
-            task.resume()
+            })
         }
-    }
 
     @available(macOS, introduced: 10.15, deprecated: 12, message: "Extension is no longer necessary. Use API built into SDK")
     @available(iOS, introduced: 13, deprecated: 15, message: "Extension is no longer necessary. Use API built into SDK")
@@ -58,16 +75,24 @@ extension URLSession {
         for request: URLRequest,
         from bodyData: Data,
         delegate: URLSessionTaskDelegate? = nil) async throws -> (Data, URLResponse) {
-        try await withCheckedThrowingContinuation { continuation in
-            let task = self.uploadTask(with: request, from: bodyData) { data, response, error in
-                guard let data = data, let response = response else {
-                    let error = error ?? URLError(.badServerResponse)
-                    return continuation.resume(throwing: error)
+            typealias ContentContinuation = CheckedContinuation<(Data, URLResponse), Error>
+            let progressWrapper = ProgressWrapper()
+            return try await withTaskCancellationHandler(handler: {
+                progressWrapper.cancel()
+            }, operation: {
+                try await withCheckedThrowingContinuation { (continuation: ContentContinuation) in
+                    let task = self.uploadTask(with: request, from: bodyData) { data, response, error in
+                        guard let data = data, let response = response else {
+                            let error = error ?? URLError(.badServerResponse)
+                            return continuation.resume(throwing: error)
+                        }
+                        
+                        continuation.resume(returning: (data, response))
+                    }
+                    progressWrapper.progress = task.progress
+                    
+                    task.resume()
                 }
-                
-                continuation.resume(returning: (data, response))
-            }
-            task.resume()
+            })
         }
-    }
 }
